Camada Controls: Adicionado a classe UpdateCategoryMethods, para que a mesma receba os métodos de seu respectivo form.
