+++ QrAguas/QrAguas.csproj
@@ -98,6 +98,7 @@
     <Compile Include="Controls\LoginMethods.cs" />
     <Compile Include="Controls\Functions.cs" />
     <Compile Include="Controls\RegisterCategoryMethods.cs" />
+    <Compile Include="Controls\UpdateCategoryMethods.cs" />
     <Compile Include="Models\CepResponse.cs" />
     <Compile Include="Models\ICepApiService.cs" />
     <Compile Include="Models\NewProduct.cs" />
